Xcos MVC: slight improvements in Adapters
 * Allow consecutive initialization thanks to data protection in DiagramAdapter.cpp
 * Allow negative values for 'from' and 'to' block field Link properties
 * Give dummy properties to Cpr and State Adapters to be able to use them, returned as TLists
 * Copy a Block's 'doc_content' field when inserting it in a diagram
 * is_valid() now logically treats link_t items
 * Regenerate ref test files

+++ scilab/modules/scicos/includes/adapters_utilities.hxx
@@ -21,8 +21,8 @@ enum startOrEnd
 
 struct link_t
 {
-    size_t block;
-    size_t port;
+    int block;
+    int port;
     enum startOrEnd kind;
 };
